+++ vscode/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.7] - 2021-11-24
+### Changed
+- Removed deprecated configuration options.
+- Renamed `toit.Path` configuration to `toit.path`.
+- Added `toitLanguageServer.command` configuration.
+
 ## [1.3.5] - 2021-11-01
 ### Changed
 - Renamed organization to project in README.